requirements.txt
chromadb==0.5.0
deepgram-sdk==3.2.7
numpy==1.26.4
ollama==0.1.9
python-dotenv==1.0.1
setuptools==69.5.1
sounddevice==0.4.6
soundfile==0.12.1
webrtcvad==2.0.10
pyautogen[teachable]==0.2.27